refactor(senpi-task): give dag amendHistory its real element type
AmendRecord depends only on DagNodeId, which types.ts already defines, so the snapshot can name the type it carries instead of widening to unknown[]. manager.ts re-exports it so no importer moves.

Plan: .omo/plans/dag-node-resume-steer.md
